peer_support/models/response_model.py
from peer_support.models import User
from django.db import models
from.question_model import Question

class Response(models.Model):
    """model used for responses"""

    user = models.ForeignKey(User, null=False, on_delete=models.CASCADE)
    question = models.ForeignKey(Question, null=False, on_delete=models.CASCADE, related_name='responses')
    parent = models.ForeignKey('self', null=True, blank=True, on_delete=models.CASCADE)
    body = models.TextField(null=False)
    created_at = models.DateTimeField(auto_now_add=True)
    updated_at = models.DateTimeField(auto_now=True)

    def get_responses(self):
        return Response.objects.filter(parent=self)